+++ src/test/java/edu/lotto/controller/AutomaticLottoTest.java
@@ -1,4 +1,4 @@
-package edu.lotto.automatic.controller;
+package edu.lotto.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
